Update- and delete- function added

+++ src/components/HelloWorld.vue
@@ -1,56 +1,90 @@
 <template>
     <div class="container">
         <div class="addTask">
-            <input v-model="nameField" placeholder="Enter title" type="text" ref="nameInput">
-            <input v-model="descriptionField" placeholder="Enter description" ref="descriptionInput">
-            <input v-model="deadlineField" placeholder="Select date" type="text" ref="deadlineInput">
-            <button type="button" @click="submitForm">
-                <font-awesome-icon icon="circle-plus" size="3x" class="circle-plus-icon" />
+            <input v-model="nameField" placeholder="Enter title" type="text" ref="nameInput" class="form-control">
+            <input v-model="descriptionField" placeholder="Enter description" ref="descriptionInput" class="form-control">
+            <input v-model="deadlineField" placeholder="Select date" type="text" ref="deadlineInput" class="form-control">
+            <button type="button" @click="submitForm" class="btn btn-primary">
+                <font-awesome-icon icon="circle-plus" size="2x" class="circle-plus-icon" />
             </button>
         </div>
         <div>
-            <div class = "errorText"></div>
+            <div class="errorText"></div>
         </div>
         <div>
-            <h3 class = "h_yourTasks">Your tasks</h3>
+            <h4 class="h_yourTasks">Your tasks</h4>
         </div>
         <div class="filter">
-            <select class="todoOptions">
+            <select class="form-select">
                 <option value="active">Active</option>
                 <option value="completed">Completed</option>
-                <option value="deleted">Deleted</option>
+                <option value="expired">Expired</option>
             </select>
         </div>
         <div class="todoTable">
-            <div class = "row">
-            </div>
+            <div class="row"></div>
             <table class="table">
                 <thead>
                 <tr>
-                    <th class ="column_checkbox"></th>
-                    <th class ="column_task">Task</th>
-                    <th class ="column_description">Description</th>
-                    <th class ="column_deadline">Deadline</th>
-                    <th class ="column_status">Status</th>
-                    <th class ="column_action"></th>
+                    <th class="column_checkbox"></th>
+                    <th class="column_task">Task</th>
+                    <th class="column_description">Description</th>
+                    <th class="column_deadline">Deadline</th>
+                    <th class="column_status">Status</th>
+                    <th class="column_action"></th>
                 </tr>
                 </thead>
-                <tbody class="tasks" data-todo-option="active">
+                <tbody class="tasks">
                 <tr v-for="task in tasks" :key="task.id">
-                    <td><input type="checkbox"></td>
-                    <td>{{ task.name }}</td>
-                    <td>{{ task.description }}</td>
-                    <td>{{ formatDate(task.deadline) }}</td>
-                    <td>{{ task.status }}</td>
-                    <td>
-                        <button @click="edit">
-                            <font-awesome-icon icon="pen" class="pen"/>
-                        </button>
-                        <button @click="remove()">
-                            <font-awesome-icon icon="trash" class="trash"/>
+                    <td class="column_checkbox">
+                        <button class="btn btn-primary-check">
+                            <font-awesome-icon icon="check" class="check" />
                         </button>
                     </td>
-
+                    <td class="column_task">
+                        <span v-if="!task.editing">{{ task.name }}</span>
+                        <input
+                            v-else
+                            type="text"
+                            v-model="task.name"
+                            class="editable-field edited-field"
+                        />
+                    </td>
+                    <td class="column_description">
+                        <span v-if="!task.editing">{{ task.description }}</span>
+                        <input
+                            v-else
+                            type="text"
+                            v-model="task.description"
+                            class="editable-field edited-field"
+                        />
+                    </td>
+                    <td class="column_deadline">
+                        <span v-if="!task.editing">{{ formatDate(task.deadline) }}</span>
+                        <input
+                            v-else
+                            type="text"
+                            v-model="task.deadline"
+                            :ref="'deadlineInput_' + task.id"
+                            class="editable-field edited-field"
+                            @focus="openDatePicker(task.id)"
+                        />
+                    </td>
+                    <td class="column_status">
+                        <span v-if="task.completed" class="text-success">Completed</span>
+                        <span v-else class="text-warning"></span>
+                    </td>
+                    <td class="column_action">
+                        <button @click="edit(task)" class="btn btn-primary" v-if="!task.editing">
+                            <font-awesome-icon icon="pen" class="pen" />
+                        </button>
+                        <button @click="update(task)" class="btn btn-success" v-if="task.editing">
+                            <font-awesome-icon icon="check" class="check" />
+                        </button>
+                        <button @click="remove(task.id)" class="btn btn-danger">
+                            <font-awesome-icon icon="trash" class="trash" />
+                        </button>
+                    </td>
                 </tr>
                 </tbody>
             </table>
@@ -66,9 +100,11 @@ import { library } from '@fortawesome/fontawesome-svg-core';
 import { faCirclePlus } from '@fortawesome/free-solid-svg-icons';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
 import { faPen } from '@fortawesome/free-solid-svg-icons';
+import { faCheck } from '@fortawesome/free-solid-svg-icons';
 library.add(faCirclePlus);
 library.add(faTrash);
 library.add(faPen);
+library.add(faCheck);
 
 export default {
     name: 'HelloWorld',
@@ -105,11 +141,14 @@ export default {
             fetch(endpoint, requestOptions)
                 .then(response => response.json())
                 .then(result => {
-                    this.tasks = result
+                    this.tasks = result.map(task => ({
+                        ...task,
+                        editing: false
+                    }));
                 })
                 .catch(error => console.log('error', error))
         },
-        save () {
+        save() {
             const endpoint = 'http://localhost:8080/add'
             const data = {
                 name: this.nameField,
@@ -127,28 +166,63 @@ export default {
                 .then(response => response.json())
                 .then(data => {
                     console.log('Success:', data)
+                    location.reload();
                 })
                 .catch(error => console.log('error', error))
         },
-        remove(id){
-            const endpoint = `http://localhost:8080/delete/${id}`;
+        remove(id) {
+            const task = this.tasks.find((task) => task.id === id);
+            if (confirm(`Are you sure you want to delete the task "${task.name}"?`)) {
+                const endpoint = `http://localhost:8080/delete/${id}`;
+                const requestOptions = {
+                    method: 'DELETE',
+                    redirect: 'follow'
+                };
+                fetch(endpoint, requestOptions)
+                    .then(result => {
+                        console.log('Deleted', result);
+                        location.reload();
+                    })
+                    .catch(error => {
+                        console.log('Error:', error);
+                    });
+            }
+        },
+        edit(task) {
+            task.editing = true;
+        },
+        update(task) {
+            const endpoint = `http://localhost:8080/edit/${task.id}`;
+            if (
+                this.nameField === task.name &&
+                this.descriptionField === task.description &&
+                this.deadlineField === task.deadline
+            ) {
+                task.editing = false;
+                return;
+            }
+            const data = {
+                name: task.name,
+                description: task.description,
+                deadline: task.deadline,
+            };
+
             const requestOptions = {
-                method: 'DELETE',
-                redirect: 'follow'
+                method: 'PUT',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(data)
             };
 
             fetch(endpoint, requestOptions)
                 .then(response => response.json())
-                .then(result => {
-                    console.log('Deleted', result);
+                .then(data => {
+                    console.log('Success:', data);
+                    this.loadTasks();
                 })
-                .catch(error => {
-                    console.log('Error:', error);
-                });
-
-        },
-        edit(){
-
+                .catch(error => console.log('error', error));
+            task.editing = false;
         },
         formatDate(date) {
             const options = {
@@ -159,8 +233,8 @@ export default {
             return new Date(date).toLocaleDateString('en-GB', options);
         },
         submitForm() {
-            let name = this.nameField.trim();
-            let date = this.deadlineField.trim();
+            let name = this.nameField;
+            let date = this.deadlineField;
             let errorText = "";
             if (name === "" && date === "") {
                 errorText = "Enter the title and the date";
@@ -174,12 +248,21 @@ export default {
             } else {
                 this.save();
                 this.loadTasks();
-                this.nameField = ""; // Reset input fields
+                this.nameField = "";
                 this.descriptionField = "";
                 this.deadlineField = "";
                 document.querySelector(".errorText").innerHTML = "";
             }
-        }
+        },
+        openDatePicker(taskId) {
+            const inputElement = this.$refs['deadlineInput_' + taskId];
+            flatpickr(inputElement, {
+                minDate: "today",
+                onChange: function (selectedDates) {
+                    this.tasks.find(task => task.id === taskId).deadline = selectedDates[0];
+                }.bind(this)
+            });
+        },
     }
 }
 
@@ -187,30 +270,54 @@ export default {
 
 <style scoped>
 
-input{
+.editable-field {
+    border: none;
+    outline: none;
+    background-color: transparent;
+    color: inherit;
+    padding: 0;
+    margin: 0;
+    font-size: inherit;
+    width: 100%;
+}
+
+.edited-field {
+    border-bottom: 1px solid gray;
+}
+
+input {
     margin-right: 25px;
-    width:100%;
+    width: 100%;
     color: #00144A;
 }
-button{
-    width:60px;
+
+button {
+    width: 50px;
+    height: 50px;
     color: #0057d2;
-    background:  #EBF8FF;
+    background: #EBF8FF;
     border-radius: 50%;
     margin-right: 20px;
-
 }
+
+.circle-plus-icon {
+    margin-left: -3px;
+    margin-top: 3px;
+}
+
 .addTask {
     width: 100%;
     display: flex;
     flex-direction: row;
     padding: 30px;
 }
+
 .container {
     max-width: 100%;
     max-height: 100%;
     color: #00144A;
 }
+
 .todoTable {
     display: flex;
     flex-direction: column;
@@ -218,22 +325,28 @@ button{
     margin-top: 50px;
     width: 100%;
 }
-.column_checkbox{
+
+.column_checkbox {
     width: 5%;
 }
-.column_task{
+
+.column_task {
     width: 20%;
 }
-.column_description{
+
+.column_description {
     width: 30%;
 }
-.column_deadline{
+
+.column_deadline {
     width: 15%;
 }
-.column_status{
+
+.column_status {
     width: 15%;
 }
-.column_action{
+
+.column_action {
     width: 15%;
 }
 
@@ -241,27 +354,40 @@ button{
     float: right;
     margin-right: 20px;
 }
+
 .todoOptions {
     width: 150px;
     color: #EBF8FF;
     background: #0057d2;
     border-radius: 30px 30px 30px 30px / 200% 200%;
 }
-.errorText{
+
+.errorText {
     color: red;
     padding-bottom: 50px;
 }
-.table{
+
+.table {
     color: #D1EFFF;
     text-align: left;
-
 }
-.h_yourTasks{
+
+.h_yourTasks {
     padding-top: 20px;
     text-align: center;
     text-transform: uppercase;
-    -webkit-text-stroke: 1px  #0057d2;
+    -webkit-text-stroke: 1px #0057d2;
     -webkit-text-fill-color: #D1EFFF;
 }
 
+.btn-primary-check:hover,
+.btn-primary-check:focus {
+    color: #28a745;
+    border-color: #28a745;
+}
+
+.btn.btn-primary-check:hover .circle-plus-icon,
+.btn.btn-primary-check:focus .circle-plus-icon {
+    color: #28a745;
+}
 </style>
